Rename isProductionDeployment
